pyproject.toml
[project]
name = "quito"
version = "0.1.0"
description = "Multi-AI quality pipeline: spec → code → review → visual QA → bugbash"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "openai>=1.30.0",
    "google-genai>=1.0.0",
    "playwright>=1.49.0",
    "pydantic>=2.10.0",
    "click>=8.1.0",
    "rich>=13.9.0",
]

[project.scripts]
quito = "quito.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
